requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==4.2.4
decorator==4.4.2
Django==5.0
django-ajax-selects==3.0.1
django-form-admin==0.5.1
django-su==1.0.0
django-unicorn==0.57.1
django-utils-six==2.0
gunicorn==21.2.0
orjson==3.9.10
packaging==23.2
Pillow==10.1.0
shortuuid==1.0.11
soupsieve==2.5
sqlparse==0.4.4
whitenoise==6.6.0
